empty files and error removed
