+++ libs/ptg-ui-angular-lib/.storybook/main.js
@@ -10,7 +10,7 @@ module.exports = {
     '../**/*.stories.mdx',
     '../**/*.stories.@(js|jsx|ts|tsx)',
   ],
-  addons: [...rootMain.addons,],
+  addons: [...rootMain.addons,'@storybook/addon-docs','@storybook/addon-controls'],
   webpackFinal: async (config, { configType }) => {
     // apply any global webpack configs that might have been specified in .storybook/main.js
     if (rootMain.webpackFinal) {
