+++ libs/ptg-ui-angular-lib/accordian/accordian/accordian.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
 
 import * as AccordianComponents from './accordian.stories.ts'
 
-import AccordianComponent from './accordian.component';
+import {AccordianComponent} from './accordian.component';
 
  <Meta title="Component/AccordianComponent/AccordianComponentDocs" component={ AccordianComponent }    />
