+++ libs/ptg-ui-angular-lib/accordian/accordian/accordian.ts
@@ -1,0 +1,19 @@
+export const ACCORDIAN_DATA = [
+    {
+      title: 'Static Header 1',
+      discription: 'This content is straight in the template.',
+    },
+    {
+      title: 'Static Header 2',
+      discription: 'This content is straight in the template.',
+    },
+    {
+      title: 'Static Header 3',
+      discription: 'This content is straight in the template.',
+    },
+    {
+      title: 'Static Header 4',
+      discription: 'This content is straight in the template.',
+    }
+  ];
+
