+++ libs/ptg-ui-angular-lib/alert/alert/alert.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, ArgsTable, Canvas, Preview } from '@storybook/addon-docs';
 
 import * as AlertComponents from './alert.stories.ts';
 
-import AlertComponent from './alert.component';
+import {AlertComponent} from './alert.component';
 
 <Meta
   title="Component/AlertComponent/AlertComponentDocs"
