+++ libs/ptg-ui-angular-lib/button/button/button.component.scss
@@ -1,49 +1,48 @@
 .ptg-ui-button {
-    font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
-    font-weight: 700;
-    border: 0;
-    border-radius: 3em;
-    cursor: pointer;
-    display: inline-block;
-    line-height: 1;
-  }
-  
-//   .ptg-ui-button--primary {
-//     color: white;
-//     color:darkcyan;
-//     background-color: #1ea7fd;
-//   }
+  font-family: 'Nunito Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif;
+  font-weight: 700;
+  border: 0;
+  border-radius: 3em;
+  cursor: pointer;
+  display: inline-block;
+  line-height: 1;
+}
 
-  .ptg-ui-button--secondary {
-    // color: #333;
-    color:black;
-    background-color: transparent;
-    box-shadow: rgba(0, 0, 0, 0.15) 0px 0px 0px 1px inset;
-  }
+.ptg-ui-button--primary {
+  background-color: #1ea7fd;
+}
 
-  .ptg-ui-button--small {
-    font-size: 8px;
-    color:black;
-    padding: 10px 10px;
-  }
+.ptg-ui-button--secondary {
+  // color: #333;
+  color: black;
+  background-color: transparent;
+  box-shadow: rgba(0, 0, 0, 0.15) 0px 0px 0px 1px inset;
+}
 
-  .ptg-ui-button--medium {
-    font-size: 12px;
-    color:black;
-    padding: 10px 12px;
-  }
+.ptg-ui-button--small {
+  background-color: #1ea7fd;
+  font-size: 12px;
+  padding: 4px 8px;
+  color: black;
+}
 
-  .ptg-ui-button--large {
-    font-size: 14px;
-    color:black;
-    padding: 12px 20px;
-  }
+.ptg-ui-button--medium {
+  font-size: 14px;
+  color: black;
+  background-color: #1ea7fd;
+  padding: 6px 12px;
+}
 
-  .ptg-ui-button--isDisable{
-    pointer-events: none;
-    opacity: 0.5;
-    color: black;
-    border: 1px solid black;
-  }
+.ptg-ui-button--large {
+  font-size: 16px;
+  color: black;
+  padding: 8px 16px;
+  background-color: #1ea7fd;
+}
 
-  
+.ptg-ui-button--isDisable {
+  pointer-events: none;
+  opacity: 0.5;
+  color: black;
+  border: 1px solid black;
+}
