+++ libs/ptg-ui-angular-lib/button/button/button.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
 
 import * as ButtonComponents from './button.stories.ts'
 
-import ButtonComponent from './button.component';
+import {ButtonComponent} from './button.component';
 
  <Meta title="Component/ButtonComponent/ButtonComponentDocs" component={ ButtonComponent }    />
