+++ libs/ptg-ui-angular-lib/calendar/calendar/calendar.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
 
 import * as CalendarComponents from './calendar.stories.ts'
 
-import CalendarComponent from './calendar.component';
+import {CalendarComponent} from './calendar.component';
 
  <Meta title="Component/CalendarComponent/CalendarComponentDocs" component={ CalendarComponent }    />
