+++ libs/ptg-ui-angular-lib/charts/d3-charts/bar-chart/bar-chart.ts
@@ -1,0 +1,24 @@
+export const BAR_CHART_D3:any = {
+    height:330,
+    width:550,
+    title:"My Charts",
+    source:'Trending languages',
+    x_title:'Languages',
+    y_title:'Numbers',
+    margin:50,
+    start: 0,
+    end: 160000,
+    data: [
+      { Framework: 'Vuee', Stars: '100000', Released: '2014', color: '#454545' },
+      { Framework: 'React', Stars: '120793', Released: '2013', color: 'blue' },
+      { Framework: 'Angular', Stars: '92342', Released: '2016', color: 'green' },
+      {
+        Framework: 'Backbone',
+        Stars: '67647',
+        Released: '2010',
+        color: 'orange',
+      },
+      { Framework: 'Ember', Stars: '121471', Released: '2011', color: 'pink' },
+    ],
+  }
+
