+++ libs/ptg-ui-angular-lib/charts/d3-charts/line-chart/line-chart.stories.ts
@@ -1,27 +1,28 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { LineChartComponent } from './line-chart.component';
-
+import {LINE_CHART_D3} from './line-chart'
 import {ChartModule} from '../../chart.module';
-export const LINE_CHART_D3:any =
-{
-  data:[
-      {
-        date:new Date("2022-03-01"),
-        value:130
-      },
-      {
-        date:new Date("2022-02-15"),
-        value:300
-      },
-      {
-        date: new Date("2022-02-01"),
-        value:1000
-      },
-  ],
-  margin: { top: 20, right: 20, bottom: 50, left: 70 },
-  width: 860,
-  height: 400
-};
+
+// export const LINE_CHART_D3:any =
+// {
+//   data:[
+//       {
+//         date:new Date("2022-03-01"),
+//         value:130
+//       },
+//       {
+//         date:new Date("2022-02-15"),
+//         value:300
+//       },
+//       {
+//         date: new Date("2022-02-01"),
+//         value:1000
+//       },
+//   ],
+//   margin: { top: 20, right: 20, bottom: 50, left: 70 },
+//   width: 860,
+//   height: 400
+// };
 export default {
   title: 'Component/LineChartComponent',
   component: LineChartComponent,
