+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/high-bar-chart/high-bar-chart.stories.ts
@@ -1,15 +1,17 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { HighBarChartComponent } from './high-bar-chart.component';
 import {ChartModule} from '../../../chart.module';
-export const BAR_CHART_3D = {
-  data: [
-    {
-      name: 'Year 1800',
-      data: [107, 120, 635, 203, 300, 203, 300]
-    }
-  ],
-  categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania', 'Canada', 'Uk']
-}
+import {BAR_CHART_3D} from './high-bar-chart'
+
+// export const BAR_CHART_3D = {
+//   data: [
+//     {
+//       name: 'Year 1800',
+//       data: [107, 120, 635, 203, 300, 203, 300]
+//     }
+//   ],
+//   categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania', 'Canada', 'Uk']
+// }
 
 export default {
   title: 'Component/HighBarChartComponent',
