+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/high-bar-chart/high-bar-chart.ts
@@ -1,0 +1,10 @@
+export const BAR_CHART_3D = {
+    data: [
+      {
+        name: 'Year 1800',
+        data: [107, 120, 635, 203, 300, 203, 300]
+      }
+    ],
+    categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania', 'Canada', 'Uk']
+  }
+
