+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/high-line-chart/high-line-chart.stories.ts
@@ -1,16 +1,9 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { HighLineChartComponent } from './high-line-chart.component';
-
+import {LINE_CHART_2D} from './high-line-chart'
 import {ChartModule} from '../../../chart.module';
-export const LINE_CHART_2D:any = {
-  data: [
-    {
-      name: 'Year 1800',
-      data: [7.0, 6.9, 9.5, 14.5, 18.2]
-    }
-  ],
-  categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
-}
+
+
 export default {
   title: 'Component/HighLineChartComponent',
   component: HighLineChartComponent,
