+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/high-line-chart/high-line-chart.ts
@@ -1,0 +1,9 @@
+export const LINE_CHART_2D:any = {
+    data: [
+      {
+        name: 'Year 1800',
+        data: [7.0, 6.9, 9.5, 14.5, 18.2]
+      }
+    ],
+    categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
+  }
