+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/high-pie-chart/high-pie-chart.stories.ts
@@ -1,22 +1,7 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { HighPieChartComponent } from './high-pie-chart.component';
-import {ChartModule} from '../../../chart.module';
-export const PIE_CHART_3D:any =
-{
-  data:[
-    ['Firefox', 45.0],
-    ['IE', 26.8],
-    {
-        name: 'Chrome',
-        y: 12.8,
-        sliced: true,
-        selected: true
-    },
-    ['Safari', 8.5],
-    ['Opera', 6.2],
-    ['Others', 0.7]
-  ]
-}
+import { ChartModule } from '../../../chart.module';
+import { PIE_CHART_3D } from './high-pie-chart';
 
 export default {
   title: 'Component/HighPieChartComponent',
@@ -37,5 +22,5 @@ const Template: Story<HighPieChartComponent> = (
 export const Primary = Template.bind({});
 Primary.args = {
   colors: ['#242582', '#8D8741', '#659DBD', '#BC986A', '#FBEEC1'],
-  data:PIE_CHART_3D.data
+  data: PIE_CHART_3D.data,
 };
