+++ libs/ptg-ui-angular-lib/charts/highcharts/2d/stacked-column-chart/stacked-column-chart.ts
@@ -1,0 +1,17 @@
+export const STACKED_COLUMN_DATA:any ={
+    title:"Stacked Column",
+    categories: ['Apples', 'Oranges', 'Pears', 'Grapes', 'Bananas'],
+    remainingOptions:{
+      series: [{
+        name: 'John',
+        data: [5, 3, 4, 7, 2]
+    }, {
+        name: 'Jane',
+        data: [2, 2, 3, 2, 1]
+    }, {
+        name: 'Joe',
+        data: [3, 4, 4, 2, 5]
+    }]
+    }
+  }
+
