+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-line-chart/high-3d-line-chart.stories.ts
@@ -1,16 +1,17 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { High3dLineChartComponent } from './high-3d-line-chart.component';
 import {ChartModule} from '../../../chart.module';
+import {LINE_CHART_2D} from './high-3d-line-chart'
 
-export const LINE_CHART_2D:any = {
-  data: [
-    {
-      name: 'Year 1800',
-      data: [7.0, 6.9, 9.5, 14.5, 18.2]
-    }
-  ],
-  categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
-}
+// export const LINE_CHART_2D:any = {
+//   data: [
+//     {
+//       name: 'Year 1800',
+//       data: [7.0, 6.9, 9.5, 14.5, 18.2]
+//     }
+//   ],
+//   categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
+// }
 
 export default {
   title: 'Component/High3dLineChartComponent',
