+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-pie-chart/high-3d-pie-chart.ts
@@ -1,0 +1,16 @@
+export const PIE_CHART_3D:any =
+  {
+    data:[
+      ['Firefox', 45.0],
+      ['IE', 26.8],
+      {
+          name: 'Chrome',
+          y: 12.8,
+          sliced: true,
+          selected: true
+      },
+      ['Safari', 8.5],
+      ['Opera', 6.2],
+      ['Others', 0.7]
+    ]
+  }
