+++ libs/ptg-ui-angular-lib/file-downloads/file-download/file-download.stories.ts
@@ -1,22 +1,8 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { FileDownloadComponent } from './file-download.component';
 import { FileDownloadsModule } from '../file-downloads.module';
+import {TABLE_DATA} from './file-download';
 
-export const TABLE_DATA = {
-  columnsData:['SN', 'NAME', 'DESIGNATION', 'DEPARTMENT'],
-  rowsData: [
-    [1, 'Astik', 'Manager', 'Engineering and MBA'],
-    [2, 'Yogita', 'QA', 'Testing'],
-    [3, 'Rajesh', 'Manager', 'Management'],
-    [4, 'Swapnil', 'Developer', 'Development'],
-    [5, 'Anmol', 'Consultant', 'HR'],
-    [6, 'Mansi', 'Software Developer', 'Engineering'],
-    [7, 'Astik', 'Manager', 'Engineering and MBA'],
-    [8, 'Yogita', 'QA', 'Testing'],
-    [9, 'Rajesh', 'Manager', 'Management'],
-    [10, 'Yogita', 'QA', 'Testing'],
-  ]
-};
 export default {
   title: 'Component/FileDownloadComponent',
   component: FileDownloadComponent,
