+++ libs/ptg-ui-angular-lib/file-downloads/file-download/file-download.ts
@@ -1,0 +1,15 @@
+export const TABLE_DATA = {
+    columnsData:['SN', 'NAME', 'DESIGNATION', 'DEPARTMENT'],
+    rowsData: [
+      [1, 'Astik', 'Manager', 'Engineering and MBA'],
+      [2, 'Yogita', 'QA', 'Testing'],
+      [3, 'Rajesh', 'Manager', 'Management'],
+      [4, 'Swapnil', 'Developer', 'Development'],
+      [5, 'Anmol', 'Consultant', 'HR'],
+      [6, 'Mansi', 'Software Developer', 'Engineering'],
+      [7, 'Astik', 'Manager', 'Engineering and MBA'],
+      [8, 'Yogita', 'QA', 'Testing'],
+      [9, 'Rajesh', 'Manager', 'Management'],
+      [10, 'Yogita', 'QA', 'Testing'],
+    ]
+  };
