+++ libs/ptg-ui-angular-lib/select/select/select.stories.ts
@@ -1,57 +1,61 @@
 import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { SelectComponent } from './select.component';
-
+import { CITY_LIST } from './selectData';
 import { SelectModule } from '../select.module'
-export const CITY_LIST = [
-  {
-    id: '1',
-    name: 'indore',
-  },
-  {
-    id: '2',
-    name: 'bhopal',
-  },
-  {
-    id: '3',
-    name: 'neemuch',
-  },
-  {
-    id: '4',
-    name: 'pune',
-  },
-  {
-    id: '5',
-    name: 'bangalore',
-  },
-  {
-    id: '6',
-    name: 'nagpur',
-  },
-  {
-    id: '7',
-    name: 'delhi',
-  },
-  {
-    id: '8',
-    name: 'hyderabad',
-  },
-  {
-    id: '9',
-    name: 'ahmedabad',
-  },
-  {
-    id: '10',
-    name: 'kanpur',
-  },
-  {
-    id: '11',
-    name: 'agra',
-  },
-  {
-    id: '12',
-    name: 'nashik',
-  }
-];
+
+// export const CITY_LIST = [
+//   {
+//     id: '1',
+//     name: 'indore',
+//   },
+//   {
+//     id: '2',
+//     name: 'bhopal',
+//   },
+//   {
+//     id: '3',
+//     name: 'neemuch',
+//   },
+//   {
+//     id: '4',
+//     name: 'pune',
+//   },
+//   {
+//     id: '5',
+//     name: 'bangalore',
+//   },
+//   {
+//     id: '6',
+//     name: 'nagpur',
+//   },
+//   {
+//     id: '7',
+//     name: 'delhi',
+//   },
+//   {
+//     id: '8',
+//     name: 'hyderabad',
+//   },
+//   {
+//     id: '9',
+//     name: 'ahmedabad',
+//   },
+//   {
+//     id: '10',
+//     name: 'kanpur',
+//   },
+//   {
+//     id: '11',
+//     name: 'agra',
+//   },
+//   {
+//     id: '12',
+//     name: 'nashik',
+//   }
+// ];
+
+
+
 export default {
   title: 'Component/SelectComponent',
   component: SelectComponent,
