+++ libs/ptg-ui-angular-lib/select/select/selectData.ts
@@ -1,0 +1,50 @@
+export const CITY_LIST = [
+    {
+      id: '1',
+      name: 'indore',
+    },
+    {
+      id: '2',
+      name: 'bhopal',
+    },
+    {
+      id: '3',
+      name: 'neemuch',
+    },
+    {
+      id: '4',
+      name: 'pune',
+    },
+    {
+      id: '5',
+      name: 'bangalore',
+    },
+    {
+      id: '6',
+      name: 'nagpur',
+    },
+    {
+      id: '7',
+      name: 'delhi',
+    },
+    {
+      id: '8',
+      name: 'hyderabad',
+    },
+    {
+      id: '9',
+      name: 'ahmedabad',
+    },
+    {
+      id: '10',
+      name: 'kanpur',
+    },
+    {
+      id: '11',
+      name: 'agra',
+    },
+    {
+      id: '12',
+      name: 'nashik',
+    }
+  ];
