+++ libs/ptg-ui-angular-lib/sidenav/sidenav/sidenav.component.html
@@ -6,7 +6,7 @@
     </button>
 
 </div>
-<!-- <div class="sidenav-container pt-3" [ngClass]="{'width-0': isMenuOpen}">
+<div class="sidenav-container pt-3" [ngClass]="{'width-0': isMenuOpen}">
     <div class="sidenav">
         <ng-container *ngFor="let item of menuItems">
             <a class="menu-item" (click)="navigate(item?.pages ? null : item)" [ngClass]="{
@@ -28,7 +28,7 @@
         </ng-container>
         <ng-content select=".app-template-name"></ng-content>
     </div>
-</div> -->
+</div>
 
 <div class="sidenav-container pt-3" [ngClass]="{'width-0': isMenuOpen}">
   <div class="sidenav">
